Moved warnAssignmentSideEffects back to Util
